fix:DB교차 검증 후 누락, 오작성 부분 수정
User
 Entity, Repository, Service 수정 보류
 Service create부분 일부 수정

+++ src/main/java/ezen/team/ezenbookstore/entity/Book.java
@@ -56,12 +56,12 @@ public class Book {
     private Integer subcategoryId;
 
     @Column(name = "count")
-    private Integer count;
+    private Long count;
 
     @Builder
     public void Builder(Long id, String title, String author, String publisher, Timestamp publishDate, Integer isbn,
                             Integer stock, String imagePath, Byte ifkr, Integer price, Integer categoryId,
-                            Integer subcategoryId, Integer count) {
+                            Integer subcategoryId, Long count) {
         this.id = id;
         this.title = title;
         this.author = author;
